requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.26.1
Pillow==10.1.0
python-dotenv==1.0.0
